renamed ollama.js to serviceInterface.js as it handles chat functionality regardless of service.
