+++ interintel.config.js
@@ -4,7 +4,7 @@ const config = {
   apiKey: `${process.env.OPENAI_API_KEY}`,
   aiService: 'ollama',
   aiVersion: `mistral`,
-  filePaths: ['./functions/openai-functions.js', './README.md'],
+  filePaths: ['ollama.js', './README.md'],
 };
 
 module.exports = config;
